requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
mysql-connector-python==8.0.26 